move view components to right folder
