+++ locallib.php
@@ -26,15 +26,11 @@
  */
 
 use core_user\fields;
-use mod_moodleoverflow\anonymous;
 use mod_moodleoverflow\capabilities;
-use mod_moodleoverflow\event\post_deleted;
 use mod_moodleoverflow\models\discussion;
-use mod_moodleoverflow\output\discussion_card;
+use mod_moodleoverflow\output\pages\view\discussion_card;
 use mod_moodleoverflow\ratings;
 use mod_moodleoverflow\readtracking;
-use mod_moodleoverflow\review;
-use mod_moodleoverflow\subscriptions;
 
 defined('MOODLE_INTERNAL') || die();
