+++ templates/pages/view/discussion_card.mustache
