+++ templates/pages/view/discussions.mustache
@@ -34,7 +34,7 @@
 
 {{#hasdiscussions}}
     {{! List of moodleoverflow where discussion can be moved to}}
-    {{> mod_moodleoverflow/view/topicmove_list}}
+    {{> mod_moodleoverflow/pages/view/topicmove_list}}
 
     {{! Paging bar if needed}}
     {{{paging_bar}}}
